feat: HomeViewController - lastFetchError 분기 처리
